Tests Passing - Resolved no collection to drop

+++ test/test_helper.js
@@ -18,21 +18,17 @@ beforeEach(done => {
     // Interesting ES6 construct
     const { ninjas } = mongoose.connection.collections
 
-    ninjas.drop()
-        // Index will be dropped so needs to be recreated
-        .then(() => { ninjas.createIndex({ 'geometry.coordinates': '2dsphere' }) })
-
-        .then(() => done())
-
-        // The first time the test is run, there is no collection to drop
-        /**
-         * REVISIT: Idea was catch here with done followed by test execution.
-         * However - Get "TypeError: Cannot read property 'drop' of undefined"
-         * Makes sense! ninjas not defined so no call to drop
-         */
-        .catch(() => done())
-        
-        // Ready to run the next text
+    // The first time the test is run, there is no collection to drop
+    if (!ninjas) {
+        done()
+    } else {
+        ninjas.drop()
+            // Index will be dropped so needs to be recreated
+            .then(() => { ninjas.createIndex({ 'geometry.coordinates': '2dsphere' }) })
+            .then(() => done())
+            .catch(() => done())
+            
+            // Ready to run the next text
+    }
 })
-
